refactor launchCommand
